Adding files management

+++ controllers/admin.js
@@ -1,5 +1,8 @@
 const mongoose = require('mongoose');
 
+const fileHelper = require('../util/file');
+
+
 const Product = require('../models/product');
 const {validationResult} = require('express-validator/check')
 
@@ -21,9 +24,25 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
+  if(!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/edit-product',
+      editing: false,
+      hasError: true,
+
+      product: {
+        title: title,
+        price: price,
+        description: description
+      },
+      errorMessage: 'Attached file is not an image',
+      validationErrors: []
+    });
+  }
   const errors = validationResult(req);
 
   if(!errors.isEmpty()) {
@@ -36,7 +55,6 @@ exports.postAddProduct = (req, res, next) => {
 
       product: {
         title: title,
-        imageUrl: imageUrl,
         price: price,
         description: description
       },
@@ -45,6 +63,9 @@ exports.postAddProduct = (req, res, next) => {
     });
 
   }
+const imageUrl = image.path;
+
+
   const product = new Product({
     // _id: mongoose.Types.ObjectId('5badf72403fd8b5e0366e81'),
     title: title,
@@ -103,8 +124,9 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
+
 
   const errors = validationResult(req);
 
@@ -118,7 +140,7 @@ exports.postEditProduct = (req, res, next) => {
 
       product: {
         title:updatedTitle,
-        imageUrl: updatedImageUrl,
+       
         price: updatedPrice,
         description: updatedDesc,
         _id: prodId
@@ -136,6 +158,12 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
+      if(image) {
+
+        fileHelper.deleteFile(product.imageUrl)
+        product.imageUrl = image.path;
+
+      }
       product.imageUrl = updatedImageUrl;
       return product.save().then(result => {
         console.log('UPDATED PRODUCT!');
@@ -170,14 +198,20 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({_id: prodId, userId: req.user._id})
-    .then(() => {
-      console.log('DESTROYED PRODUCT');
-      res.redirect('/admin/products');
-    })
-    .catch(err => {
-      const error  = new Error(err);
-       error.httpStatusCode = 500;
-       return next(error);
-     });
+  Product.findById(prodId).then(product => {
+    if(!product){
+      return next(new Error('Product not found.'));
+    }
+    fileHelper.deleteFile(product.imageUrl)
+    return Product.deleteOne({_id: prodId, userId: req.user._id})
+  }) .then(() => {
+    console.log('DESTROYED PRODUCT');
+    res.redirect('/admin/products');
+  })
+  .catch(err => {
+    const error  = new Error(err);
+     error.httpStatusCode = 500;
+     return next(error);
+   });
+   
 };
